refactor(tui): the naming rule holds with no exception, and the tenants keep their own state
The renderer package took the plain name and the plugin took the suffix, which is the one place the naming rule did not transliterate directory, package and id. packages/tui is packages/tui-renderer now and the plugin takes the plain name.

The choices a command opened and the questions that stand keep their own state, each in its own module, so a second ask no longer overwrites the first and the surface holds neither map.
