+++ apps/cli/src/attach.test.ts
@@ -25,7 +25,7 @@ import { toolPolicy } from "@missingstudio/eva-tool-policy"
 import { toolRead } from "@missingstudio/eva-tool-read"
 import { trace } from "@missingstudio/eva-trace"
 import { traceMemory } from "@missingstudio/eva-trace-memory"
-import { ASKING, makeSurface, TUI_SURFACE, type SurfaceDeps } from "@missingstudio/eva-tui-surface"
+import { ASKING, makeSurface, TUI_SURFACE, type SurfaceDeps } from "@missingstudio/eva-tui"
 import { makeWeb, WEB_SURFACE } from "@missingstudio/eva-web"
 import { Effect, Exit, Scope } from "effect"
 import { describe, expect, it } from "vitest"
