+++ packages/conformance/package.json
@@ -45,7 +45,7 @@
     "@missingstudio/eva-trace-sqlite": "workspace:*",
     "@missingstudio/eva-tui": "workspace:*",
     "@missingstudio/eva-tui-core": "workspace:*",
-    "@missingstudio/eva-tui-surface": "workspace:*",
+    "@missingstudio/eva-tui-renderer": "workspace:*",
     "@missingstudio/eva-validator": "workspace:*",
     "@missingstudio/eva-web": "workspace:*",
     "@missingstudio/eva-web-app": "workspace:*",
