+++ packages/tui-renderer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@missingstudio/eva-tui",
+  "name": "@missingstudio/eva-tui-renderer",
   "version": "0.0.0",
   "private": true,
   "type": "module",
