+++ packages/tui-renderer/src/app.tsx
