+++ packages/tui-renderer/src/frame.test.ts
