+++ packages/tui-renderer/src/frame.ts
