+++ packages/tui-renderer/src/palette.ts
