+++ packages/tui-renderer/src/render-check.tsx
