+++ packages/tui-renderer/src/renderer.tsx
