+++ packages/tui-renderer/src/stream.test.ts
