+++ packages/tui-renderer/src/stream.ts
