+++ plugins/tui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@missingstudio/eva-tui-surface",
+  "name": "@missingstudio/eva-tui",
   "version": "0.0.0",
   "private": true,
   "type": "module",
